Ing de Datos 2/TP/seed_service/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
redis
pymongo
cassandra-driver
neo4j